moved gtk/gtksourceview to gtksourceview for consistency
