+++ Makefile
@@ -8,7 +8,7 @@ all:
 	cd gdk && gomake
 	cd gdkpixbuf && gomake
 	cd gtk && gomake
-	cd gtk/gtksourceview && gomake
+	cd gtksourceview && gomake
 
 install:
 	cd pango && gomake install
@@ -16,7 +16,7 @@ install:
 	cd gdk && gomake install
 	cd gdkpixbuf && gomake install
 	cd gtk && gomake install
-	cd gtk/gtksourceview && gomake install
+	cd gtksourceview && gomake install
 
 clean:
 	cd pango && gomake clean
@@ -24,18 +24,18 @@ clean:
 	cd gdk && gomake clean
 	cd gdkpixbuf && gomake clean
 	cd gtk && gomake clean
-	cd gtk/gtksourceview && gomake clean
+	cd gtksourceview && gomake clean
 	cd example && gomake clean
 
 fmt_all:
 	gofmt ./gdk/gdk.go  > ./gdk/gdk.go.fmt
 	gofmt ./gtk/gtk.go > ./gtk/gtk.go.fmt
-	gofmt ./gtk/gtksourceview/gtksourceview.go > ./gtk/gtksourceview/gtksourceview.go.fmt
+	gofmt ./gtksourceview/gtksourceview.go > .gtksourceview/gtksourceview.go.fmt
 	gofmt ./gdkpixbuf/gdkpixbuf.go > ./gdkpixbuf/gdkpixbuf.go.fmt
 	gofmt ./glib/glib.go > ./glib/glib.go.fmt
 	gofmt ./pango/pango.go > ./pango/pango.go.fmt
 	mv ./gtk/gtk.go.fmt ./gtk/gtk.go
-	mv ./gtk/gtksourceview/gtksourceview.go.fmt ./gtk/gtksourceview/gtksourceview.go
+	mv ./gtksourceview/gtksourceview.go.fmt ./gtksourceview/gtksourceview.go
 	mv ./gdk/gdk.go.fmt ./gdk/gdk.go
 	mv ./gdkpixbuf/gdkpixbuf.go.fmt ./gdkpixbuf/gdkpixbuf.go
 	mv ./glib/glib.go.fmt ./glib/glib.go
